fix lint dependecy error
