+++ packages/runtime/src/widgets/Charts/BarChart.tsx
@@ -2,7 +2,8 @@ import { Bar } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
 import { useRegisterBindings } from "@ensembleui/react-framework";
 import { useState } from "react";
-import { getMergedOptions, type ChartDataSets, type ChartProps } from "..";
+import { type ChartDataSets, type ChartProps } from "..";
+import { getMergedOptions } from "./utils/getMergedOptions";
 
 const options: ChartOptions<"bar"> = {
   maintainAspectRatio: false,
