+++ packages/runtime/src/widgets/Charts/LineChart.tsx
@@ -2,8 +2,9 @@ import { Line } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
 import { useState } from "react";
 import { useRegisterBindings } from "@ensembleui/react-framework";
-import { getMergedOptions, type ChartDataSets, type ChartProps } from "..";
+import { type ChartDataSets, type ChartProps } from "..";
 import { get } from "lodash-es";
+import { getMergedOptions } from "./utils/getMergedOptions";
 
 const options: ChartOptions<"line"> = {
   maintainAspectRatio: false,
