+++ packages/runtime/src/widgets/Charts/StackBarChart.tsx
@@ -2,9 +2,10 @@ import { Bar } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
 import { useRegisterBindings } from "@ensembleui/react-framework";
 import { useState } from "react";
-import { getMergedOptions, type ChartDataSets, type ChartProps } from "..";
+import { type ChartDataSets, type ChartProps } from "..";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import { get } from "lodash-es";
+import { getMergedOptions } from "./utils/getMergedOptions";
 
 const options: ChartOptions<"bar"> = {
   maintainAspectRatio: false,
